% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLabelSystem.R
\name{getLabelSystem}
\alias{getLabelSystem}
\title{getLabelSystem}
\usage{
getLabelSystem(systemName)
}
\arguments{
\item{systemName}{the labeling system info to return
\itemize{
  \item mindboggle
  \item mindboggle-extended
  \item antsct
 }}
}
\description{
get info for a labeling scheme
}
