% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuants.R
\name{getQuants}
\alias{getQuants}
\title{getQuants}
\usage{
getQuants(path, id, date = NULL, system = NULL, label = NULL,
  measure = NULL, metric = NULL, as.wide = FALSE,
  with.filenames = F)
}
\arguments{
\item{path}{base directory for ANTsCT output}

\item{id}{list of subjects to collect values for}

\item{date}{optional list of dates for each subject (one-to-one mapping with id list)}

\item{system}{labeling systems to obtain values for}

\item{label}{label numbers of interest}

\item{measure}{what measure/s to collect}

\item{metric}{which metric/s to collect}

\item{as.wide}{flag to return wide format data frame}
}
\description{
get values of interest from pre-generated .csv files
}
