% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeBilateralLabels.R
\name{mergeBilateralLabels}
\alias{mergeBilateralLabels}
\title{mergeBilateralLabels}
\usage{
mergeBilateralLabels(labels, labelSystem = "mindboggle")
}
\arguments{
\item{labels}{the image of labels or its filename}

\item{labelSystem}{scheme for labels, only "mindboggle" currently supported}
}
\description{
get a mask created by merging specific labels
}
