% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindBoggleLabels.R
\docType{data}
\name{mindBoggleLabels}
\alias{mindBoggleLabels}
\title{mindBoggleLabels}
\format{A data frame with 201 rows and 8 variables:
\itemize{
  \item number: number used in labeling of image
  \item common: does this label typically appear in the data
  \item name: full name of region
  \item abbreviation: abbreviated name of region
  \item hemisphere: right or left
  \item lobe: lobe groupings
  \item cortex: is this in cortex
  \item bilateral: label grouping together bilateral regions
 }}
\usage{
mindBoggleLabels
}
\description{
A dataset containing mind boggle labels and associated information:
}
\examples{
data(mindBoggleLabels)
}
\keyword{datasets}
