% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantsRename.R
\name{quantsRename}
\alias{quantsRename}
\title{quantsRename}
\usage{
quantsRename(input, renaming, keep.measure = F, keep.metric = F,
  seperator = "_")
}
\arguments{
\item{input}{data.frame with column names to replace ( i.e. antsct_1_volume_numeric ) or vector of column names}

\item{renaming}{a data.frame with columns 'name' and 'number' or a csv filename}

\item{keep.measure}{append the measure name in the output name}

\item{keep.metric}{append the metric name in the output name}

\item{seperator}{character used when appending meaure and metric}
}
\description{
change the naming for quants columns
}
