% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceQuants.R
\name{reduceQuants}
\alias{reduceQuants}
\title{reduceQuants}
\usage{
reduceQuants(x, name, measure, metric)
}
\arguments{
\item{x}{data.frame obtained from getQuants() that contains volume and one metric/measure (e.g. mean thickness)}

\item{name}{what to call the reduce value}

\item{measure}{the measure to reduce (e.g. "thickness")}

\item{metric}{the metric to reduce (e.g. "mean")}
}
\description{
merge values from multiple label ( e.g. merge all frontal regions )
}
