% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectLabels.R
\name{subjectLabels}
\alias{subjectLabels}
\title{subjectLabels}
\usage{
subjectLabels(labels, image = NULL, mask = NULL, outfile = NULL)
}
\arguments{
\item{labels}{the image of labels or its filename}

\item{image}{the image whose values are summarized, or its filename}

\item{mask}{an image mask to apply to labels, or it's filename}

\item{outfile}{filename for resulting data to be saved}
}
\description{
summarize a subjects labels:
}
