% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuantsLabelNumbers.R
\name{getQuantsLabelNumbers}
\alias{getQuantsLabelNumbers}
\title{getQuantsLabelNumbers}
\usage{
getQuantsLabelNumbers(system = "mindboggle", cortex = NA, group = NA,
  hemisphere = NA, return.index = FALSE)
}
\arguments{
\item{system}{labeling systems to obtain values for}
}
\description{
get label numbers for regions of interest
}
