% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mindBoggleLabels.R
\docType{data}
\name{mindBoggleLabels}
\alias{mindBoggleLabels}
\title{mindBoggleLabels}
\format{A data frame with 201 rows and 8 variables:
\itemize{
  \item Number: number used in labeling of image
  \item Common: does this label typically appear in the data
  \item Name: full name of region
  \item Abbreviation: abbreviated name of region
  \item Hemisphere: right or left
  \item Lobe: lobe groupings
  \item Cortex: is this in cortex
  \item Bilateral: label grouping together bilateral regions
}}
\usage{
mindBoggleLabels
}
\description{
A dataset containing mind boggle labels and associated information:
}
\keyword{datasets}
