% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectLabelStats.R
\name{subjectLabelStats}
\alias{subjectLabelStats}
\title{subjectLabelStats}
\usage{
subjectLabelStats(labels, image = NULL, mask = NULL, weights = NULL,
  outfile = NULL, labelSet = NULL, labelSystem = "mindboggle")
}
\arguments{
\item{labels}{the image of labels or its filename}

\item{image}{the image whose values are summarized, or its filename}

\item{mask}{an image mask to apply to labels, or it's filename}

\item{outfile}{filename for resulting data to be saved}

\item{labelSet}{an array of all valid label values}
}
\description{
summarize a subjects labels:
}
